Conaco/src/requirements.txt
fastapi
uvicorn
pillow
supabase
requests
python-dotenv
